backend/requirements.txt
# Main dependencies
asgiref==3.8.1
build==1.2.2.post1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pip-tools==7.4.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
sqlparse==0.5.3
uritemplate==4.1.1
wheel==0.45.1

# Testing dependencies
coverage==7.8.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
